fix: Change stageProgress from Int to Double to fix JSON parsing crash
The API returns stageProgress as a floating-point number (e.g., 0.33495145631068),
but it was declared as Int which caused JSON deserialization to fail when
encountering the decimal point.

Changes:
- ReadaloudApiModel: stageProgress Int? -> Double?
- ReadaloudDomainModel: stageProgress Int? -> Double?
- ReadaloudLocalModel: stageProgress Int? -> Double?
- ReadaloudEntity: stageProgress Int? -> Double?
- ReadaloudSqlDelightEntity: stageProgress Int? -> Double?
- Readaloud.sq: stage_progress INTEGER -> REAL
- BooksSqlDelightDao: Remove toLong/toInt conversions for stageProgress
- Bump database schema version from 2 to 3
- Add migration file 2.sqm

+++ lib/database/implementation/src/commonMain/kotlin/com/retro99/database/implementation/dao/books/ReadaloudSqlDelightEntity.kt
@@ -12,7 +12,7 @@ data class ReadaloudSqlDelightEntity(
     override val missing: Int?,
     override val status: String?,
     override val currentStage: String?,
-    override val stageProgress: Int?,
+    override val stageProgress: Double?,
     override val queuePosition: Int?,
     override val restartPending: Boolean?,
     override val createdAt: String?,
